+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This document outlines major changes between releases.
 - Context cancellation in DeleteMultipleObjects handler leading to pool client status issues (#1212)
 - Useless "search old object versions failed" warning for new objects (#1211)
 - Duplicates in directory listings in some cases (#1213)
+- Multipart upload in case of more than 9 parts (#1214)
 
 ### Updated